Split comment API functions out into separate module
This basically just moves some code around, pulling all the
comment-related API functions out into a separate module.

part of #1667
pr #1764
