+++ static/js/lib/api/api.js
@@ -1,8 +1,6 @@
 // @flow
 /* global SETTINGS:false, fetch: false */
 // For mocking purposes we need to use "fetch" defined as a global instead of importing as a local.
-import R from "ramda"
-import qs from "query-string"
 import { PATCH, POST, DELETE } from "redux-hammock/constants"
 import { fetchJSONWithCSRF } from "redux-hammock/django_csrf_fetch"
 
@@ -12,13 +10,10 @@ import {
   fetchWithAuthFailure,
   fetchJSONWithToken
 } from "./fetch_auth"
-import { getCommentSortQS, getPaginationSortQS } from "./util"
+import { getPaginationSortQS } from "./util"
 
 import type { AuthResponse, AuthFlow } from "../../flow/authTypes"
 import type {
-  GenericComment,
-  CommentFromAPI,
-  MoreCommentsFromAPI,
   UserFeedComment,
   Post,
   Profile,
@@ -37,68 +32,6 @@ export function search(params: SearchParams): Promise<*> {
     method: POST,
     body:   JSON.stringify(body)
   })
-}
-
-export function getComments(
-  postID: string,
-  params: Object
-): Promise<Array<CommentFromAPI | MoreCommentsFromAPI>> {
-  return fetchJSONWithAuthFailure(
-    `/api/v0/posts/${postID}/comments/${getCommentSortQS(params)}`
-  )
-}
-
-export const getComment = (
-  commentID: string
-): Promise<Array<CommentFromAPI | MoreCommentsFromAPI>> =>
-  fetchJSONWithAuthFailure(`/api/v0/comments/${commentID}`)
-
-export const createComment = (
-  postId: string,
-  text: string,
-  commentId: ?string
-): Promise<CommentFromAPI> => {
-  const body =
-    commentId === undefined ? { text } : { text, comment_id: commentId }
-
-  return fetchJSONWithAuthFailure(`/api/v0/posts/${postId}/comments/`, {
-    method: POST,
-    body:   JSON.stringify(body)
-  })
-}
-
-export function updateComment(
-  commentId: string,
-  commentPayload: Object
-): Promise<GenericComment> {
-  return fetchJSONWithAuthFailure(`/api/v0/comments/${commentId}/`, {
-    method: PATCH,
-    body:   JSON.stringify(commentPayload)
-  })
-}
-
-export function deleteComment(commentId: string): Promise<GenericComment> {
-  return fetchWithAuthFailure(`/api/v0/comments/${commentId}/`, {
-    method: DELETE
-  })
-}
-
-export function getMoreComments(
-  postId: string,
-  parentId: string | null,
-  children: Array<string>
-): Promise<Array<CommentFromAPI | MoreCommentsFromAPI>> {
-  const payload = {
-    post_id:  postId,
-    children: children
-  }
-  if (!R.isNil(parentId)) {
-    // $FlowFixMe: Flow doesn't know that we're still constructing payload
-    payload.parent_id = parentId
-  }
-  return fetchJSONWithAuthFailure(
-    `/api/v0/morecomments/?${qs.stringify(payload)}`
-  )
 }
 
 export function getUserPosts(
