+++ static/js/lib/api/api_test.js
@@ -1,20 +1,12 @@
 /* global SETTINGS: false */
 import { assert } from "chai"
 import sinon from "sinon"
-import qs from "query-string"
-import { PATCH, POST, DELETE } from "redux-hammock/constants"
+import { PATCH, POST } from "redux-hammock/constants"
 import * as fetchFuncs from "redux-hammock/django_csrf_fetch"
-import R from "ramda"
 
 import {
-  getComments,
-  getComment,
-  createComment,
-  updateComment,
-  getMoreComments,
   getUserPosts,
   getUserComments,
-  deleteComment,
   getSettings,
   patchFrontpageSetting,
   patchCommentSetting,
@@ -34,11 +26,7 @@ import {
   patchWidgetList
 } from "./api"
 import { makePost, makeChannelPostList } from "../../factories/posts"
-import {
-  makeCommentsResponse,
-  makeMoreCommentsResponse
-} from "../../factories/comments"
-import { COMMENT_SORT_NEW } from "../picker"
+import { makeCommentsResponse } from "../../factories/comments"
 import * as authFuncs from "./fetch_auth"
 import * as searchFuncs from "../search"
 import { makeProfile } from "../../factories/profiles"
@@ -66,101 +54,6 @@ describe("api", function() {
     beforeEach(() => {
       fetchJSONStub = sandbox.stub(fetchFuncs, "fetchJSONWithCSRF")
       fetchStub = sandbox.stub(fetchFuncs, "fetchWithCSRF")
-    })
-
-    describe("getComments", () => {
-      let post, response
-
-      beforeEach(() => {
-        post = makePost()
-        response = makeCommentsResponse(post)
-        fetchJSONStub.returns(Promise.resolve(response))
-      })
-
-      it("gets comments for a post", async () => {
-        const resp = await getComments(post.id, {})
-        assert.deepEqual(resp, response)
-      })
-
-      it("includes the sort parameter when getting comments", async () => {
-        const resp = await getComments(post.id, { sort: COMMENT_SORT_NEW })
-        assert.deepEqual(resp, response)
-        assert.ok(
-          fetchJSONStub.calledWith(
-            `/api/v0/posts/${post.id}/comments/?sort=new`
-          )
-        )
-      })
-    })
-
-    it("gets a single comment", async () => {
-      const post = makePost()
-      const response = R.slice(0, 1, makeCommentsResponse(post))
-      fetchJSONStub.returns(response)
-
-      const resp = await getComment(post.id)
-      assert.deepEqual(resp, response)
-    })
-
-    it("creates comments for a post", async () => {
-      const post = makePost()
-      fetchJSONStub.returns(Promise.resolve())
-
-      await createComment(post.id, "my new comment")
-      assert.ok(fetchJSONStub.calledWith(`/api/v0/posts/${post.id}/comments/`))
-      assert.deepEqual(fetchJSONStub.args[0][1], {
-        method: POST,
-        body:   JSON.stringify({ text: "my new comment" })
-      })
-    })
-
-    it("creates comments replying to comments", async () => {
-      const post = makePost()
-      const tree = makeCommentsResponse(post)
-      fetchJSONStub.returns(Promise.resolve())
-
-      await createComment(post.id, "my new comment", tree[0].id)
-      assert.ok(fetchJSONStub.calledWith(`/api/v0/posts/${post.id}/comments/`))
-      assert.deepEqual(fetchJSONStub.args[0][1], {
-        method: POST,
-        body:   JSON.stringify({
-          text:       "my new comment",
-          comment_id: tree[0].id
-        })
-      })
-    })
-
-    it("updates a comment", async () => {
-      const post = makePost()
-      const tree = makeCommentsResponse(post)
-      const comment = tree[0]
-      const commentResponse = { ...comment, replies: undefined, text: "edited" }
-
-      fetchJSONStub.returns(Promise.resolve(commentResponse))
-
-      const payload = {
-        text:      "edited",
-        downvoted: true
-      }
-      const updated = await updateComment(comment.id, payload)
-      assert.ok(fetchJSONStub.calledWith(`/api/v0/comments/${comment.id}/`))
-      assert.deepEqual(updated, commentResponse)
-      assert.deepEqual(fetchJSONStub.args[0][1], {
-        method: PATCH,
-        body:   JSON.stringify(payload)
-      })
-    })
-
-    it("deletes a comment", async () => {
-      const comment = makeCommentsResponse(makePost())[0]
-      fetchStub.returns(Promise.resolve())
-
-      await deleteComment(comment.id)
-      assert.ok(
-        fetchStub.calledWith(`/api/v0/comments/${comment.id}/`, {
-          method: DELETE
-        })
-      )
     })
 
     it("gets profile", async () => {
@@ -216,51 +109,6 @@ describe("api", function() {
           .then(() => {
             checkArgs()
           })
-      })
-    })
-
-    describe("retrieves more comments", () => {
-      it("at the root level", async () => {
-        const post = makePost()
-        const moreComments = makeMoreCommentsResponse(post)
-        const children = ["some", "child", "ren"]
-
-        fetchJSONStub.returns(Promise.resolve(moreComments))
-
-        const response = await getMoreComments(post.id, null, children)
-        const payload = {
-          post_id:  post.id,
-          children: children
-        }
-        assert.ok(
-          fetchJSONStub.calledWith(
-            `/api/v0/morecomments/?${qs.stringify(payload)}`
-          )
-        )
-        assert.deepEqual(response, moreComments)
-      })
-
-      it("replying to a parent", async () => {
-        const post = makePost()
-        const commentsResponse = makeCommentsResponse(post)
-        const parent = commentsResponse[0]
-        const moreComments = makeMoreCommentsResponse(post, parent.id)
-        const children = ["some", "child", "ren"]
-
-        fetchJSONStub.returns(Promise.resolve(moreComments))
-
-        const response = await getMoreComments(post.id, parent.id, children)
-        const payload = {
-          post_id:   post.id,
-          parent_id: parent.id,
-          children:  children
-        }
-        assert.ok(
-          fetchJSONStub.calledWith(
-            `/api/v0/morecomments/?${qs.stringify(payload)}`
-          )
-        )
-        assert.deepEqual(response, moreComments)
       })
     })
